RMIO-77 Upgrade sharpziplib to 1.4.2
Remove obsolete test code

Was used to create a nonfunctional test case for the umlaut in name test if no implementation provided. The singleton which enabled the wrong behaviour is now obsolete, which also means this case should not occur anymore, rendering these lines useless.  The global mechanism still does not create a red test here, meaning this code is redundant.